[MOV] website_event_track_live: move public widgets to interactions folder
