+++ addons/website_event_track_live/static/src/interactions/website_event_track_replay_suggestion.js
