+++ addons/website_event_track_live_quiz/__manifest__.py
@@ -20,7 +20,6 @@
     'assets': {
         'web.assets_frontend': [
             'website_event_track_live_quiz/static/src/interactions/**/*',
-            'website_event_track_live_quiz/static/src/js/**/*',
             'website_event_track_live_quiz/static/src/xml/**/*',
         ],
     },
